requirements.txt
# pip install -r requirements.txt
langchain==0.1.13
chainlit
transformers==4.39.1
huggingface-hub==0.21.4
langchain-community==0.0.29
python-dotenv==1.0.1
fastapi==0.110.0
# torch
# InstructorEmbedding
# sentence-transformers
# faiss-cpu
# python-dotenv
# altair
# tiktoken